+++ tests/PhpSpreadsheetTests/Reader/Xlsx/ChartsTitleTest.php
@@ -1,23 +1,24 @@
 <?php
 
-namespace PhpOffice\PhpSpreadsheetTests\Reader;
+namespace PhpOffice\PhpSpreadsheetTests\Reader\Xlsx;
 
+use PhpOffice\PhpSpreadsheet\Chart\Title;
 use PhpOffice\PhpSpreadsheet\IOFactory;
 use PHPUnit\Framework\TestCase;
 
-function getTitleText($title)
-{
-    if (null === $title || null === $title->getCaption()) {
-        return null;
-    }
-
-    return implode("\n", array_map(function ($rt) {
-        return $rt->getPlainText();
-    }, $title->getCaption()));
-}
-
 class ChartsTitleTest extends TestCase
 {
+    private static function getTitleText(?Title $title): ?string
+    {
+        if (null === $title || empty($title->getCaption())) {
+            return null;
+        }
+
+        return implode("\n", array_map(function ($rt) {
+            return $rt->getPlainText();
+        }, $title->getCaption())); // @phpstan-ignore-line
+    }
+
     public function testChartTitles(): void
     {
         $filename = 'tests/data/Reader/XLSX/excelChartsTest.xlsx';
@@ -31,34 +32,34 @@ class ChartsTitleTest extends TestCase
 
         // No title or axis labels
         $chart1 = $charts[0];
-        $title = getTitleText($chart1->getTitle());
+        $title = self::getTitleText($chart1->getTitle());
         self::assertEmpty($title);
-        self::assertEmpty(getTitleText($chart1->getXAxisLabel()));
-        self::assertEmpty(getTitleText($chart1->getYAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart1->getXAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart1->getYAxisLabel()));
 
         // Title, no axis labels
         $chart2 = $charts[1];
 
-        self::assertEquals('Chart with Title and no Axis Labels', getTitleText($chart2->getTitle()));
-        self::assertEmpty(getTitleText($chart2->getXAxisLabel()));
-        self::assertEmpty(getTitleText($chart2->getYAxisLabel()));
+        self::assertEquals('Chart with Title and no Axis Labels', self::getTitleText($chart2->getTitle()));
+        self::assertEmpty(self::getTitleText($chart2->getXAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart2->getYAxisLabel()));
 
         // No title, only horizontal axis label
         $chart3 = $charts[2];
-        self::assertEmpty(getTitleText($chart3->getTitle()));
-        self::assertEquals('Horizontal Axis Title Only', getTitleText($chart3->getXAxisLabel()));
-        self::assertEmpty(getTitleText($chart3->getYAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart3->getTitle()));
+        self::assertEquals('Horizontal Axis Title Only', self::getTitleText($chart3->getXAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart3->getYAxisLabel()));
 
         // No title, only vertical axis label
         $chart4 = $charts[3];
-        self::assertEmpty(getTitleText($chart4->getTitle()));
-        self::assertEquals('Vertical Axis Title Only', getTitleText($chart4->getYAxisLabel()));
-        self::assertEmpty(getTitleText($chart4->getXAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart4->getTitle()));
+        self::assertEquals('Vertical Axis Title Only', self::getTitleText($chart4->getYAxisLabel()));
+        self::assertEmpty(self::getTitleText($chart4->getXAxisLabel()));
 
         // Title and both axis labels
         $chart5 = $charts[4];
-        self::assertEquals('Complete Annotations', getTitleText($chart5->getTitle()));
-        self::assertEquals('Horizontal Axis Title', getTitleText($chart5->getXAxisLabel()));
-        self::assertEquals('Vertical Axis Title', getTitleText($chart5->getYAxisLabel()));
+        self::assertEquals('Complete Annotations', self::getTitleText($chart5->getTitle()));
+        self::assertEquals('Horizontal Axis Title', self::getTitleText($chart5->getXAxisLabel()));
+        self::assertEquals('Vertical Axis Title', self::getTitleText($chart5->getYAxisLabel()));
     }
 }
